python-ocr-service/requirements.txt
fastapi
paddleocr
pillow
uvicorn
